+++ test/integration/express_route_circuit.rb
@@ -30,7 +30,7 @@ resources.resource_groups.create(
   name: 'TestRG-ER',
   location: 'eastus'
 )
-
+puts 'Resource Group created!'
 ########################################################################################################################
 ################                               Create Express Route Circuit                            #################
 ########################################################################################################################
@@ -60,6 +60,7 @@ network.express_route_circuits.create(
     }
   ]
 )
+puts 'Express Route Circuit created!'
 
 ########################################################################################################################
 ######################                           List Express Route Circuit                         ####################
@@ -71,10 +72,48 @@ circuits.each do |circuit|
 end
 
 ########################################################################################################################
-######################                                   CleanUp                                  ######################
+######################                Create Express Route Circuit Authorizations                   ####################
 ########################################################################################################################
+network.express_route_circuit_authorizations.create(
+  resource_group: 'TestRG-ER',
+  circuit_name: 'testERCircuit',
+  authorization_status: 'Available',
+  authorization_name: 'Test-Auth',
+  name: 'Unique-Auth-Name'
+)
+puts 'Express Route Circuit Authorization created!'
 
+########################################################################################################################
+######################                Get a Express Route Circuit Authorization                     ####################
+########################################################################################################################
+authorization = network.express_route_circuit_authorizations.get('TestRG-ER', 'testERCircuit', 'Test-Auth')
+puts authorization.name.to_s
+
+########################################################################################################################
+######################                List Express Route Circuit Authorizations                     ####################
+########################################################################################################################
+authorizations = network.express_route_circuit_authorizations(resource_group: 'TestRG-ER', circuit_name: 'testERCircuit')
+authorizations.each do |auth|
+  puts auth.name.to_s
+end
+
+########################################################################################################################
+######################                Destroy Express Route Circuit Authorization                   ####################
+########################################################################################################################
+authorization = network.express_route_circuit_authorizations.get('TestRG-ER', 'testERCircuit', 'Test-Auth')
+authorization.destroy
+puts 'Express Route Circuit Authorization deleted!'
+
+########################################################################################################################
+######################                   Destroy Express Route Circuit                            ######################
+########################################################################################################################
 circuit = network.express_route_circuits.get('TestRG-ER', 'testERCircuit')
-puts circuit.name
-
 circuit.destroy
+puts 'Express Route Circuit deleted!'
+
+########################################################################################################################
+######################                         CleanUp                                            ######################
+########################################################################################################################
+resource_group = resources.resource_groups.get('TestRG-ER')
+resource_group.destroy
+puts 'Resource Group deleted!'